+++ .eslintrc.js
@@ -5,4 +5,9 @@ module.exports = {
     extends: [
         'pandora-typescript',
     ],
+    "settings": {
+        "react": {
+            "version": "16.0",
+        },
+    },
 };