+++ lib/signature-extend.ts
@@ -1,15 +1,15 @@
 import _ from 'lodash';
 import Crypto from 'client-crypto';
 import uuid from 'uuid/v4';
-import { isFormData } from '../utils/form';
-import { IAjax, IAjaxProcessDataOptions, IParams, IMethod, IOptions } from '../interface';
+import { isFormData } from './utils/form';
+import { IAjax, IAjaxProcessDataOptions, IParams, IMethod, IOptions } from './interface';
 
 /**
- * 签名扩展
- * 将会在请求头中添加字段sign，timestamp，app-nonce
- * sign：签名文本
- * timestamp（签名参数）：UTC时间（用于校验是否已过期）
- * app-nonce（签名参数）：只使用一次标识码（用于校验是否已发送过，存入redis几分钟后过期）
+ * 签名扩展。
+ * 将会在请求头中添加字段sign，timestamp，app-nonce。
+ * sign：签名文本；
+ * timestamp（签名参数）：UTC时间（用于校验是否已过期）；
+ * app-nonce（签名参数）：只使用一次标识码（用于校验是否已发送过，存入redis几分钟后过期）。
  */
 function signatureExtend(): () => void {
     return function signature(): void {